requirements.txt
Flask==3.0.2
Werkzeug==3.0.1
requests==2.31.0
pycryptodome==3.20.0
protobuf==4.25.1
